+++ MO.Servers/MO.Grains/Game/PlayerData.cs
@@ -5,14 +5,14 @@ namespace MO.Grains.Game
     public class PlayerData
     {
         private IUser _user;
-        public double X { get; set; }
-        public double Y { get; set; }
+        public float X { get; set; }
+        public float Y { get; set; }
         public PlayerData(IUser user)
         {
             _user = user;
         }
 
-        public void SetPoint(double x, double y)
+        public void SetPoint(float x, float y)
         {
             X = x;
             Y = y;
